requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpaca-trade-api==2.3.0
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
Cython==0.29.31
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.9.1
fastapi==0.79.0
fastjsonschema==2.16.1
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
idna==2.10
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
multidict==6.0.2
multitasking==0.0.11
mysql-connector-python==8.0.29
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
newtulipy==0.4.6
notebook==6.4.12
numerize==0.12
numpy==1.20.3
openpyxl==3.0.7
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
polygon-api-client==1.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyarrow==9.0.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
rich==12.5.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
stack-data==0.3.0
starlette==0.19.1
streamlit==1.11.1
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
unicorn==2.0.0
urllib3==1.26.11
uvicorn==0.18.2
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
widgetsnbextension==3.6.1
xlrd==2.0.1
yarg==0.1.9
yarl==1.7.2
yfinance==0.1.74
zipp==3.8.1
